requirements.txt
django~=5.0.0
djangorestframework
django-cors-headers
python-dateutil
Pillow
gunicorn
mysqlclient
whitenoise
